Rest of i64 -> i32

+++ crates/db_schema/src/source/combined/search.rs
@@ -27,7 +27,7 @@ use serde_with::skip_serializing_none;
 pub struct SearchCombined {
   pub id: SearchCombinedId,
   pub published_at: DateTime<Utc>,
-  pub score: i64,
+  pub score: i32,
   pub post_id: Option<PostId>,
   pub comment_id: Option<CommentId>,
   pub community_id: Option<CommunityId>,
